refactored deepwounds and dotuptime into Typescript and statistic bar for debuff (same as shadow priest dots)
